server/requirements.txt
fastapi[standard]
python_dotenv
uvicorn